+++ scripts/winesapos-install.sh
@@ -1068,16 +1068,6 @@ if [[ "${WINESAPOS_BUILD_CHROOT_ONLY}" == "false" ]]; then
     # Configure support for older Intel iGPUs.
     sed -i s'/GRUB_CMDLINE_LINUX="/GRUB_CMDLINE_LINUX="i915.force_probe="*" /'g ${WINESAPOS_INSTALL_DIR}/etc/default/grub
 
-    # Configure Arch Linux and SteamOS to load the Linux kernels in the correct order of newest to oldest.
-    # This will make the newest kernel be bootable by default. For example, on Arch Linux 'linux' will be
-    # the default over 'linux-lts' and on SteamOS 'linux-lts' will be the default over 'linux-steamos'.
-    # https://github.com/LukeShortCloud/winesapOS/issues/144
-    # https://github.com/LukeShortCloud/winesapOS/issues/325
-    # https://github.com/LukeShortCloud/winesapOS/issues/489
-    if [[ "${WINESAPOS_DISTRO_DETECTED}" != "manjaro" ]]; then
-        sed -i s'/version_sort\ -r/sort/'g ${WINESAPOS_INSTALL_DIR}/etc/grub.d/10_linux
-    fi
-
     efi_partition=2
     if [[ "${WINESAPOS_ENABLE_PORTABLE_STORAGE}" == "true" ]]; then
         efi_partition=3