+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0 - The Conversion Update
 
 - Remove no longer needed "amdgpu.sg_display=0" workaround.
+- Change 'linux-t2' to be the default Linux kernel instead of 'linux-lts'.
 
 ## 3.4.0 - The Game Mode Update